Slave informs master about abi/platform/timeout
This commit should cut down on the number of duplicate builds at the end
of the queue by tracking which builds are currently "active". It also
expands the details sent from the slave to the master which we can use
in the next version to request abi-specific builds.

Also sort out some of the logging so it's a little less verbose on the
master when --verbose is used.