Fix #23
Remove high_priest and move all the task handling logic back into
master's main as it's much simpler to coordinate shutdown when there's
not "two mains" as it were.  Ditch the PUB/SUB internal control queue;
the tasks have to start in a particular order and they need to shut down
in a particular order so each needs its own (inproc) control queue to
time individual QUIT messages.

Enhanced the interaction and appearance of the monitor; it now receives
an immediate update of all slaves on startup, and displays slave ID,
ABI, and overall uptime in the UI.