Workaround: #31
Ensure that we create directories and empty indexes even for builds that
fail. Related to upstream bug pypa/pip#4580. If a user attempts to build
a wheel for a package with `--no-cache-dir` that exists on PyPI, but not
on piwheels (and all builds have failed on piwheels, so no directory
exists at all for that package), then the build fails. Creating a dummy
directory on piwheels alleviates this.

Not strictly our issue, but upstream is still open and even if they
fixed it tomorrow there's no guarantee it'd appear in Raspbian (or
Debian upstream) quickly, plus the workaround is pretty simple (although
some surgery to the file-server will be needed to create dummy dirs and
indexes for all existing failed builds).