Fix up db queries and re-work terminal module
This is a squash of a bunch of commits because I'm horribly rushing a
bit rebase onto master before I have to head out!

This fixes up the various build-related things in the database to cope
with including the ABI in potential builds, re-works the terminal module
so it's a simple bunch of functions, splits out all the default queue
names into a const module, and finishes off the initdb script.