Parse wheel for package & version
Also reject imports for non-existent packages. Given the disparity
between package names and wheel filenames (- for _ and so on) it's
awfully easy to specify the wrong name. Hence we now extract and parse
the metadata.json for the *real* package name, and reject any overrides
which don't exist. Might extend the import protocol to allow forcing
registration in future...