Ensure no dead-locks during shutdown
While stopping tasks, the main thread wasn't continuing to drain the
internal status queue. If a task was blocked sending to the status queue
and the main thread asked the task to terminate a dead-lock ensued. Also
extend the stats queue which is occasionally blocking with the extra
build slaves in place.