Fix #127
Only log the build in the database once all files from the build have
been successfully transferred. This prevents the (rare, but still
possible) circumstance that a successful build is logged but the files
never get transferred (e.g. because the master restarts immediately
after) resulting in an erroneous index entry next time the index is
updated.

This fix does lead to the possibility of a file being transferred then
the master restarting and the corresponding build never being logged but
that's less of a problem: firstly the file won't appear in an index
(because the build wasn't logged) and secondly another slave will
subsequently redo the build (because the build wasn't logged).