Add SLEEP and WAKE commands to master control
Along with all the necessary plumbing to pass this to slave-driver and
store the corresponding state. This should (eventually) enable the
following controls from the monitor: kill everything immediately, wait
for all jobs to finish then kill everything, pause everything
immediately, wait for all jobs to finish then pause everything, and
per-slave variants of all the above (plus skip the current job on
an individual slave).