Re #198
Never download binary packages from PyPI; always build (I think this
option may not have existed in older pip versions but it does now, and
we "pip install --upgrade pip" on the slaves so we ought to have it).

This isn't a full fix for the issue (he says wording things carefully to
avoid GitHub prematurely closing the ticket!) hence the title, but it's
important enough to warrant its own commit.