Re-factor screen transitions
Rather than two writable attributes, use a single method to update both
renderer and transition simultaneously.

At this point, it's also become obvious that a lot of code is shared
between the urwid and sense-hat monitors in terms of state tracking.
Rather than continue the duplicate, the next commit will be a refactor
of the two applications under a common module.