Start tasks synchronously
Wait a while for task ready on current task before starting the next
(but if the task isn't ready after 10 seconds, carry on starting tasks
anyway). This avoids a (very rare) race condition between the-scribe and
file-juggler where the latter tries to query the state of the output
directory before the former has created it.