Simplify recent-delete handling
The architect's queue has a (minimal) depth which means the simplistic
recent_deletes set is insufficient. Better to just re-use the
recent_builds dictionary for the same purpose with an arbitrary timeout
for deleted packages/versions of an hour. Also renamed recent_builds to
excluded_builds to reflect the new purpose and added a bit of logging to
give a clue what's going on.