Fix early deletion affecting build-queue
There's a potential race right at the start of operation (which is why
this crash started almost immediately into a run) where a package or
version deletion can be registered *before* the build-queue has been
populated. In this case it was possible for a deleted package to wind up
in the queue because the excluded list was basing potential ABIs off the
abi_queue list of ABIs. Now it queries the database...