New audit script
At one point this turned into an over-engineered mess; this is a
wind-back to a much simpler version that does the basic job. Not
quickly, and not elegantly but it doesn't stress the machine being
audited (useful for running on an active index), and doesn't wind up
horribly difficult to terminate because various sub-processes are
blocked waiting on various queues all over the place.