Workaround sphinx's slightly crap mock
Sphinx's mock (while producing very nice __repr__ output for
documentation), doesn't provide any mocking of the standard operators
like | (or) so if the mocks get used in any prototypes (as they do in
piwheels.transport.Poller.register) the docs fail to build. This is a
nasty workaround but it's the only sensible thing to do for now.