Bump the schedule on big-brother
At this point, generation of the homepage takes about 5 minutes, and
the search index about 10. With the current delays this swamps the
big-brother process, which in turn delays the slave-driver process to
the point that all the builders timeout periodically.

This is a crap and temporary workaround; the proper workaround is to
have the-architect periodically run these and push the results to
big-brother (and the-scribe).