Use decorators to generate TheOracle
Rather than laboriously write out all the RPC stuff that the oracle
engages in, just use decorators to annotate the necessary methods on the
Database class with what we need to generate handlers (the message
associated with a method, and functions to convert call parameters into
serialized data, and back again).

This should *greatly* simplify future expansions of the stored proc API
in the database as we'll just need to define the stored proc itself, the
handler in the Database class, and the entry in protocols covering the
prototype of its message parameters.